Prevent Valgrind warning with unlink(NULL).
git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@1052 ea778897-0a13-0410-b9d1-a72fbfd435f5

+++ src/socket.c
@@ -160,8 +160,11 @@ gint socket_finalize(void)
 #ifdef G_OS_WIN32
 	WSACleanup();
 #else
-	unlink(socket_info.file_name);
-	g_free(socket_info.file_name);
+	if (socket_info.file_name)
+	{
+		unlink(socket_info.file_name);
+		g_free(socket_info.file_name);
+	}
 #endif
 
 	return 0;
